requirements.txt
langchain
langchain-google-genai
python-dotenv
langchain-community
pydantic
requests
openai
# langchain_anthropic
langchain_huggingface
langchain-huggingface 
huggingface-hub
google-generativeai
langchain_openai
streamlit
langchain_google_genai
langchain-groq groq









